requirements.txt
pydrive
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
python-telegram-bot==13.7.0
pyrogram
